First

+++ src/Movie.java
@@ -1,14 +1,15 @@
 public class Movie {
+// Klassen Movie repræsenter film-delen af vores MovieCollection
 
-    // Attributes
+    // ATTRIBUTTER
     private String title;
     private String director;
     private int year;
     private boolean isInColor;
-    private double lengthInMinutes;
+    private int lengthInMinutes;
     private String genre;
 
-    // Constructor
+    // KONSTRUKTØR tager attributerne som argumenter og initialiserer objektet med dem
     public Movie(String title, String director, int year, boolean isInColor, int lengthInMinutes, String genre) {
         this.title = title;
         this.director = director;
@@ -18,32 +19,17 @@ public class Movie {
         this.genre = genre;
     }
 
-    // Getters
-    public String getTitle() {
-        return title;
-    }
-
-    public String getDirector() {
-        return director;
-    }
-
-    public int getYear() {
-        return year;
-    }
-
-    public boolean getIsInColor() {
-        return isInColor;
-    }
-
-    public double getLengthInMinutes() {
-        return lengthInMinutes;
-    }
-
-    public String getGenre() {
-        return genre;
-    }
+    // GETTER giver adgang til værdigerne af attributterne
+    public String getTitle() {return title;}
+    public String getDirector() {return director;}
+    public int getYear() {return year;}
+    public boolean getIsInColor() {return isInColor;}
+    public int getLengthInMinutes() {return lengthInMinutes;}
+    public String getGenre() {return genre;}
 
     @Override
+    // METODE: toString findes i alle objekter, da den er arvet fra klassen 'Objekt' og konverterer objektets tilstand til en streng.
+    // toString returner en String repræsentation af et objekt. Bruger vi ikke toString, printer programmet volapyk ud.
     public String toString() {
         String result = "";
         result += "Title: " + title + "\n" + "Director: " + director + "\n" + "Year: " + year + "\n" + "Genre: " + genre + "\n" + "Runtime: " + lengthInMinutes + "\n";
@@ -52,8 +38,6 @@ public class Movie {
         } else {
             result += "Color: Black and White\n";
         }
-
         return result;
-
     }
 }
